llm-port-project/src/aura/ai_agency_wealth/requirements.txt
crewai
langchain-community
duckduckgo-search
python-dotenv
qrcode[pil]
fpdf2
resend
litellm